Expose the handler for consumers
